move protoc_gen_openapiv2 into api, remove ignores, remove untyped sub
