+++ temporalio/Steepfile
@@ -11,7 +11,6 @@ common_diagnostics = {
 target :lib do
   signature 'sig'
   check 'lib'
-  ignore 'lib/protoc-gen-openapiv2'
   library 'uri', 'objspace', 'etc'
   configure_code_diagnostics do |hash|
     hash.update(common_diagnostics)
