+++ temporalio/lib/temporalio/api/cloud/cloudservice/v1/service.rb
@@ -6,7 +6,7 @@ require 'google/protobuf'
 
 require 'temporalio/api/cloud/cloudservice/v1/request_response'
 require 'google/api/annotations_pb'
-require 'protoc-gen-openapiv2/options/annotations_pb'
+require 'temporalio/api/protoc_gen_openapiv2/options/annotations'
 
 
 descriptor_data = "\n0temporal/api/cloud/cloudservice/v1/service.proto\x12\"temporal.api.cloud.cloudservice.v1\x1a\x39temporal/api/cloud/cloudservice/v1/request_response.proto\x1a\x1cgoogle/api/annotations.proto\x1a.protoc-gen-openapiv2/options/annotations.proto2\xb6\xbe\x01\n\x0c\x43loudService\x12\xb0\x02\n\x12GetCurrentIdentity\x12=.temporal.api.cloud.cloudservice.v1.GetCurrentIdentityRequest\x1a>.temporal.api.cloud.cloudservice.v1.GetCurrentIdentityResponse\"\x9a\x01\x92\x41x\n\x07\x41\x63\x63ount\x12\x14Get current identity\x1aWReturns information about the currently authenticated user or service account principal\x82\xd3\xe4\x93\x02\x19\x12\x17/cloud/current-identity\x12\xa5\x02\n\x08GetUsers\x12\x33.temporal.api.cloud.cloudservice.v1.GetUsersRequest\x1a\x34.temporal.api.cloud.cloudservice.v1.GetUsersResponse\"\xad\x01\x92\x41\x95\x01\n\x05Users\x12\x0eList all users\x1a*Returns a list of all users in the account\"E\n\x1dUser management documentation\x12$https://docs.temporal.io/cloud/users*\tlistUsers\x82\xd3\xe4\x93\x02\x0e\x12\x0c/cloud/users\x12\x9c\x02\n\x07GetUser\x12\x32.temporal.api.cloud.cloudservice.v1.GetUserRequest\x1a\x33.temporal.api.cloud.cloudservice.v1.GetUserResponse\"\xa7\x01\x92\x41\x85\x01\n\x05Users\x12\x0eGet user by ID\x1a%Takes a user ID, returns user details\"E\n\x1dUser management documentation\x12$https://docs.temporal.io/cloud/users\x82\xd3\xe4\x93\x02\x18\x12\x16/cloud/users/{user_id}\x12\xd0\x01\n\nCreateUser\x12\x35.temporal.api.cloud.cloudservice.v1.CreateUserRequest\x1a\x36.temporal.api.cloud.cloudservice.v1.CreateUserResponse\"S\x92\x41\x39\n\x05Users\x12\rCreate a user\x1a!Creates a new user in the account\x82\xd3\xe4\x93\x02\x11\"\x0c/cloud/users:\x01*\x12\xdb\x01\n\nUpdateUser\x12\x35.temporal.api.cloud.cloudservice.v1.UpdateUserRequest\x1a\x36.temporal.api.cloud.cloudservice.v1.UpdateUserResponse\"^\x92\x41:\n\x05Users\x12\rUpdate a user\x1a\"Updates an existing user\'s details\x82\xd3\xe4\x93\x02\x1b\"\x16/cloud/users/{user_id}:\x01*\x12\xd5\x01\n\nDeleteUser\x12\x35.temporal.api.cloud.cloudservice.v1.DeleteUserRequest\x1a\x36.temporal.api.cloud.cloudservice.v1.DeleteUserResponse\"X\x92\x41\x37\n\x05Users\x12\rDelete a user\x1a\x1fRemoves a user from the account\x82\xd3\xe4\x93\x02\x18*\x16/cloud/users/{user_id}\x12\xaa\x03\n\x16SetUserNamespaceAccess\x12\x41.temporal.api.cloud.cloudservice.v1.SetUserNamespaceAccessRequest\x1a\x42.temporal.api.cloud.cloudservice.v1.SetUserNamespaceAccessResponse\"\x88\x02\x92\x41\xc5\x01\n\x05Users\x12\x19Set user namespace access\x1a\x38\x43onfigures a user\'s permissions for a specific namespace\"g\n#Namespace permissions documentation\x12@https://docs.temporal.io/cloud/users-namespace-level-permissions\x82\xd3\xe4\x93\x02\x39\"4/cloud/namespaces/{namespace}/users/{user_id}/access:\x01*\x12\xb1\x02\n\x11GetAsyncOperation\x12<.temporal.api.cloud.cloudservice.v1.GetAsyncOperationRequest\x1a=.temporal.api.cloud.cloudservice.v1.GetAsyncOperationResponse\"\x9e\x01\x92\x41m\n\nOperations\x12\x1aGet async operation status\x1a\x43Returns the current status and details of an asynchronous operation\x82\xd3\xe4\x93\x02(\x12&/cloud/operations/{async_operation_id}\x12\xc6\x02\n\x0f\x43reateNamespace\x12:.temporal.api.cloud.cloudservice.v1.CreateNamespaceRequest\x1a;.temporal.api.cloud.cloudservice.v1.CreateNamespaceResponse\"\xb9\x01\x92\x41\x99\x01\n\nNamespaces\x12\x12\x43reate a namespace\x1a&Creates a new namespace in the account\"O\n\"Namespace management documentation\x12)https://docs.temporal.io/cloud/namespaces\x82\xd3\xe4\x93\x02\x16\"\x11/cloud/namespaces:\x01*\x12\xc7\x02\n\rGetNamespaces\x12\x38.temporal.api.cloud.cloudservice.v1.GetNamespacesRequest\x1a\x39.temporal.api.cloud.cloudservice.v1.GetNamespacesResponse\"\xc0\x01\x92\x41\xa3\x01\n\nNamespaces\x12\x13List all namespaces\x1a/Returns a list of all namespaces in the account\"O\n\"Namespace management documentation\x12)https://docs.temporal.io/cloud/namespaces\x82\xd3\xe4\x93\x02\x13\x12\x11/cloud/namespaces\x12\xda\x02\n\x0cGetNamespace\x12\x37.temporal.api.cloud.cloudservice.v1.GetNamespaceRequest\x1a\x38.temporal.api.cloud.cloudservice.v1.GetNamespaceResponse\"\xd6\x01\x92\x41\xad\x01\n\nNamespaces\x12\x15Get namespace details\x1a\x37Returns detailed information about a specific namespace\"O\n\"Namespace management documentation\x12)https://docs.temporal.io/cloud/namespaces\x82\xd3\xe4\x93\x02\x1f\x12\x1d/cloud/namespaces/{namespace}\x12\xdb\x02\n\x0fUpdateNamespace\x12:.temporal.api.cloud.cloudservice.v1.UpdateNamespaceRequest\x1a;.temporal.api.cloud.cloudservice.v1.UpdateNamespaceResponse\"\xce\x01\x92\x41\xa2\x01\n\nNamespaces\x12\x12Update a namespace\x1a/Updates configuration for an existing namespace\"O\n\"Namespace management documentation\x12)https://docs.temporal.io/cloud/namespaces\x82\xd3\xe4\x93\x02\"\"\x1d/cloud/namespaces/{namespace}:\x01*\x12\xc7\x03\n\x1bRenameCustomSearchAttribute\x12\x46.temporal.api.cloud.cloudservice.v1.RenameCustomSearchAttributeRequest\x1aG.temporal.api.cloud.cloudservice.v1.RenameCustomSearchAttributeResponse\"\x96\x02\x92\x41\xcb\x01\n\nNamespaces\x12\x1eRename custom search attribute\x1a\x30Renames a custom search attribute in a namespace\"k\n&Custom Search Attributes documentation\x12\x41https://docs.temporal.io/search-attribute#custom-search-attribute\x82\xd3\xe4\x93\x02\x41\"</cloud/namespaces/{namespace}/rename-custom-search-attribute:\x01*\x12\xd9\x02\n\x0f\x44\x65leteNamespace\x12:.temporal.api.cloud.cloudservice.v1.DeleteNamespaceRequest\x1a;.temporal.api.cloud.cloudservice.v1.DeleteNamespaceResponse\"\xcc\x01\x92\x41\xa3\x01\n\nNamespaces\x12\x12\x44\x65lete a namespace\x1a\x30Permanently deletes a namespace and all its data\"O\n\"Namespace management documentation\x12)https://docs.temporal.io/cloud/namespaces\x82\xd3\xe4\x93\x02\x1f*\x1d/cloud/namespaces/{namespace}\x12\xaf\x03\n\x17\x46\x61iloverNamespaceRegion\x12\x42.temporal.api.cloud.cloudservice.v1.FailoverNamespaceRegionRequest\x1a\x43.temporal.api.cloud.cloudservice.v1.FailoverNamespaceRegionResponse\"\x8a\x02\x92\x41\xce\x01\n\x11High Availability\x12\x19\x46\x61ilover namespace region\x1a?Initiates a regional failover for a high availability namespace\"]\n)High availability namespace documentation\x12\x30https://docs.temporal.io/cloud/high-availability\x82\xd3\xe4\x93\x02\x32\"-/cloud/namespaces/{namespace}/failover-region:\x01*\x12\x83\x03\n\x12\x41\x64\x64NamespaceRegion\x12=.temporal.api.cloud.cloudservice.v1.AddNamespaceRegionRequest\x1a>.temporal.api.cloud.cloudservice.v1.AddNamespaceRegionResponse\"\xed\x01\x92\x41\xb6\x01\n\x11High Availability\x12\x15\x41\x64\x64 namespace replica\x1a+Adds a new replica to an existing namespace\"]\n)High availability namespace documentation\x12\x30https://docs.temporal.io/cloud/high-availability\x82\xd3\xe4\x93\x02-\"(/cloud/namespaces/{namespace}/add-region:\x01*\x12\x9b\x03\n\x15\x44\x65leteNamespaceRegion\x12@.temporal.api.cloud.cloudservice.v1.DeleteNamespaceRegionRequest\x1a\x41.temporal.api.cloud.cloudservice.v1.DeleteNamespaceRegionResponse\"\xfc\x01\x92\x41\xc2\x01\n\x11High Availability\x12\x18Remove namespace replica\x1a\x34Removes a replica from a high availability namespace\"]\n)High availability namespace documentation\x12\x30https://docs.temporal.io/cloud/high-availability\x82\xd3\xe4\x93\x02\x30*./cloud/namespaces/{namespace}/regions/{region}\x12\xa3\x02\n\nGetRegions\x12\x35.temporal.api.cloud.cloudservice.v1.GetRegionsRequest\x1a\x36.temporal.api.cloud.cloudservice.v1.GetRegionsResponse\"\xa5\x01\x92\x41\x8b\x01\n\x07Regions\x12\x10List all regions\x1a-Returns a list of all available cloud regions\"?\n\x15Regions documentation\x12&https://docs.temporal.io/cloud/regions\x82\xd3\xe4\x93\x02\x10\x12\x0e/cloud/regions\x12\xb2\x02\n\tGetRegion\x12\x34.temporal.api.cloud.cloudservice.v1.GetRegionRequest\x1a\x35.temporal.api.cloud.cloudservice.v1.GetRegionResponse\"\xb7\x01\x92\x41\x94\x01\n\x07Regions\x12\x12Get region details\x1a\x34Returns detailed information about a specific region\"?\n\x15Regions documentation\x12&https://docs.temporal.io/cloud/regions\x82\xd3\xe4\x93\x02\x19\x12\x17/cloud/regions/{region}\x12\xa8\x02\n\nGetApiKeys\x12\x35.temporal.api.cloud.cloudservice.v1.GetApiKeysRequest\x1a\x36.temporal.api.cloud.cloudservice.v1.GetApiKeysResponse\"\xaa\x01\x92\x41\x8f\x01\n\x08\x41PI Keys\x12\x11List all API keys\x1a-Returns a list of all API keys in the account\"A\n\x16\x41PI Keys documentation\x12\'https://docs.temporal.io/cloud/api-keys\x82\xd3\xe4\x93\x02\x11\x12\x0f/cloud/api-keys\x12\xb8\x02\n\tGetApiKey\x12\x34.temporal.api.cloud.cloudservice.v1.GetApiKeyRequest\x1a\x35.temporal.api.cloud.cloudservice.v1.GetApiKeyResponse\"\xbd\x01\x92\x41\x99\x01\n\x08\x41PI Keys\x12\x13Get API key details\x1a\x35Returns detailed information about a specific API key\"A\n\x16\x41PI Keys documentation\x12\'https://docs.temporal.io/cloud/api-keys\x82\xd3\xe4\x93\x02\x1a\x12\x18/cloud/api-keys/{key_id}\x12\xb1\x02\n\x0c\x43reateApiKey\x12\x37.temporal.api.cloud.cloudservice.v1.CreateApiKeyRequest\x1a\x38.temporal.api.cloud.cloudservice.v1.CreateApiKeyResponse\"\xad\x01\x92\x41\x8f\x01\n\x08\x41PI Keys\x12\x11\x43reate an API key\x1a-Creates a new API key for programmatic access\"A\n\x16\x41PI Keys documentation\x12\'https://docs.temporal.io/cloud/api-keys\x82\xd3\xe4\x93\x02\x14\"\x0f/cloud/api-keys:\x01*\x12\xb5\x02\n\x0cUpdateApiKey\x12\x37.temporal.api.cloud.cloudservice.v1.UpdateApiKeyRequest\x1a\x38.temporal.api.cloud.cloudservice.v1.UpdateApiKeyResponse\"\xb1\x01\x92\x41\x8a\x01\n\x08\x41PI Keys\x12\x11Update an API key\x1a(Updates an existing API key\'s properties\"A\n\x16\x41PI Keys documentation\x12\'https://docs.temporal.io/cloud/api-keys\x82\xd3\xe4\x93\x02\x1d\"\x18/cloud/api-keys/{key_id}:\x01*\x12\xa8\x02\n\x0c\x44\x65leteApiKey\x12\x37.temporal.api.cloud.cloudservice.v1.DeleteApiKeyRequest\x1a\x38.temporal.api.cloud.cloudservice.v1.DeleteApiKeyResponse\"\xa4\x01\x92\x41\x80\x01\n\x08\x41PI Keys\x12\x11\x44\x65lete an API key\x1a\x1eRevokes and deletes an API key\"A\n\x16\x41PI Keys documentation\x12\'https://docs.temporal.io/cloud/api-keys\x82\xd3\xe4\x93\x02\x1a*\x18/cloud/api-keys/{key_id}\x12\xc3\x02\n\x11GetNexusEndpoints\x12<.temporal.api.cloud.cloudservice.v1.GetNexusEndpointsRequest\x1a=.temporal.api.cloud.cloudservice.v1.GetNexusEndpointsResponse\"\xb0\x01\x92\x41\x8e\x01\n\x05Nexus\x12\x18List all Nexus endpoints\x1a\x34Returns a list of all Nexus endpoints in the account\"5\n\x13Nexus documentation\x12\x1ehttps://docs.temporal.io/nexus\x82\xd3\xe4\x93\x02\x18\x12\x16/cloud/nexus/endpoints\x12\xd8\x02\n\x10GetNexusEndpoint\x12;.temporal.api.cloud.cloudservice.v1.GetNexusEndpointRequest\x1a<.temporal.api.cloud.cloudservice.v1.GetNexusEndpointResponse\"\xc8\x01\x92\x41\x98\x01\n\x05Nexus\x12\x1aGet Nexus endpoint details\x1a<Returns detailed information about a specific Nexus endpoint\"5\n\x13Nexus documentation\x12\x1ehttps://docs.temporal.io/nexus\x82\xd3\xe4\x93\x02&\x12$/cloud/nexus/endpoints/{endpoint_id}\x12\xd5\x02\n\x13\x43reateNexusEndpoint\x12>.temporal.api.cloud.cloudservice.v1.CreateNexusEndpointRequest\x1a?.temporal.api.cloud.cloudservice.v1.CreateNexusEndpointResponse\"\xbc\x01\x92\x41\x97\x01\n\x05Nexus\x12\x17\x43reate a Nexus endpoint\x1a>Creates a new Nexus endpoint for cross-namespace communication\"5\n\x13Nexus documentation\x12\x1ehttps://docs.temporal.io/nexus\x82\xd3\xe4\x93\x02\x1b\"\x16/cloud/nexus/endpoints:\x01*\x12\xd7\x02\n\x13UpdateNexusEndpoint\x12>.temporal.api.cloud.cloudservice.v1.UpdateNexusEndpointRequest\x1a?.temporal.api.cloud.cloudservice.v1.UpdateNexusEndpointResponse\"\xbe\x01\x92\x41\x8b\x01\n\x05Nexus\x12\x17Update a Nexus endpoint\x1a\x32Updates an existing Nexus endpoint\'s configuration\"5\n\x13Nexus documentation\x12\x1ehttps://docs.temporal.io/nexus\x82\xd3\xe4\x93\x02)\"$/cloud/nexus/endpoints/{endpoint_id}:\x01*\x12\xcb\x02\n\x13\x44\x65leteNexusEndpoint\x12>.temporal.api.cloud.cloudservice.v1.DeleteNexusEndpointRequest\x1a?.temporal.api.cloud.cloudservice.v1.DeleteNexusEndpointResponse\"\xb2\x01\x92\x41\x82\x01\n\x05Nexus\x12\x17\x44\x65lete a Nexus endpoint\x1a)Removes a Nexus endpoint from the account\"5\n\x13Nexus documentation\x12\x1ehttps://docs.temporal.io/nexus\x82\xd3\xe4\x93\x02&*$/cloud/nexus/endpoints/{endpoint_id}\x12\xcc\x02\n\rGetUserGroups\x12\x38.temporal.api.cloud.cloudservice.v1.GetUserGroupsRequest\x1a\x39.temporal.api.cloud.cloudservice.v1.GetUserGroupsResponse\"\xc5\x01\x92\x41\xa7\x01\n\x06Groups\x12\x14List all user groups\x1a\x30Returns a list of all user groups in the account\"U\n\x19User groups documentation\x12\x38https://docs.temporal.io/cloud/users-account-level-roles\x82\xd3\xe4\x93\x02\x14\x12\x12/cloud/user-groups\x12\xd0\x02\n\x0cGetUserGroup\x12\x37.temporal.api.cloud.cloudservice.v1.GetUserGroupRequest\x1a\x38.temporal.api.cloud.cloudservice.v1.GetUserGroupResponse\"\xcc\x01\x92\x41\xa3\x01\n\x06Groups\x12\x16Get user group details\x1a\x38Returns detailed information about a specific user group\"G\n\x19User groups documentation\x12*https://docs.temporal.io/cloud/user-groups\x82\xd3\xe4\x93\x02\x1f\x12\x1d/cloud/user-groups/{group_id}\x12\xc7\x02\n\x0f\x43reateUserGroup\x12:.temporal.api.cloud.cloudservice.v1.CreateUserGroupRequest\x1a;.temporal.api.cloud.cloudservice.v1.CreateUserGroupResponse\"\xba\x01\x92\x41\x99\x01\n\x06Groups\x12\x13\x43reate a user group\x1a\x31\x43reates a new user group for managing permissions\"G\n\x19User groups documentation\x12*https://docs.temporal.io/cloud/user-groups\x82\xd3\xe4\x93\x02\x17\"\x12/cloud/user-groups:\x01*\x12\xcc\x02\n\x0fUpdateUserGroup\x12:.temporal.api.cloud.cloudservice.v1.UpdateUserGroupRequest\x1a;.temporal.api.cloud.cloudservice.v1.UpdateUserGroupResponse\"\xbf\x01\x92\x41\x93\x01\n\x06Groups\x12\x13Update a user group\x1a+Updates an existing user group\'s properties\"G\n\x19User groups documentation\x12*https://docs.temporal.io/cloud/user-groups\x82\xd3\xe4\x93\x02\"\"\x1d/cloud/user-groups/{group_id}:\x01*\x12\xc3\x02\n\x0f\x44\x65leteUserGroup\x12:.temporal.api.cloud.cloudservice.v1.DeleteUserGroupRequest\x1a;.temporal.api.cloud.cloudservice.v1.DeleteUserGroupResponse\"\xb6\x01\x92\x41\x8d\x01\n\x06Groups\x12\x13\x44\x65lete a user group\x1a%Removes a user group from the account\"G\n\x19User groups documentation\x12*https://docs.temporal.io/cloud/user-groups\x82\xd3\xe4\x93\x02\x1f*\x1d/cloud/user-groups/{group_id}\x12\xad\x03\n\x1bSetUserGroupNamespaceAccess\x12\x46.temporal.api.cloud.cloudservice.v1.SetUserGroupNamespaceAccessRequest\x1aG.temporal.api.cloud.cloudservice.v1.SetUserGroupNamespaceAccessResponse\"\xfc\x01\x92\x41\xb2\x01\n\x06Groups\x12\x1fSet user group namespace access\x1a>Configures a user group\'s permissions for a specific namespace\"G\n\x19User groups documentation\x12*https://docs.temporal.io/cloud/user-groups\x82\xd3\xe4\x93\x02@\";/cloud/namespaces/{namespace}/user-groups/{group_id}/access:\x01*\x12\xdf\x02\n\x12\x41\x64\x64UserGroupMember\x12=.temporal.api.cloud.cloudservice.v1.AddUserGroupMemberRequest\x1a>.temporal.api.cloud.cloudservice.v1.AddUserGroupMemberResponse\"\xc9\x01\x92\x41\x95\x01\n\x06Groups\x12\x11\x41\x64\x64 user to group\x1a/Adds a user to a user group (Cloud groups only)\"G\n\x19User groups documentation\x12*https://docs.temporal.io/cloud/user-groups\x82\xd3\xe4\x93\x02*\"%/cloud/user-groups/{group_id}/members:\x01*\x12\xf8\x02\n\x15RemoveUserGroupMember\x12@.temporal.api.cloud.cloudservice.v1.RemoveUserGroupMemberRequest\x1a\x41.temporal.api.cloud.cloudservice.v1.RemoveUserGroupMemberResponse\"\xd9\x01\x92\x41\x9f\x01\n\x06Groups\x12\x16Remove user from group\x1a\x34Removes a user from a user group (Cloud groups only)\"G\n\x19User groups documentation\x12*https://docs.temporal.io/cloud/user-groups\x82\xd3\xe4\x93\x02\x30\"+/cloud/user-groups/{group_id}/remove-member:\x01*\x12\xdf\x02\n\x13GetUserGroupMembers\x12>.temporal.api.cloud.cloudservice.v1.GetUserGroupMembersRequest\x1a?.temporal.api.cloud.cloudservice.v1.GetUserGroupMembersResponse\"\xc6\x01\x92\x41\x95\x01\n\x06Groups\x12\x15List users in a group\x1a+Returns a list of all users in a user group\"G\n\x19User groups documentation\x12*https://docs.temporal.io/cloud/user-groups\x82\xd3\xe4\x93\x02\'\x12%/cloud/user-groups/{group_id}/members\x12\xf5\x02\n\x14\x43reateServiceAccount\x12?.temporal.api.cloud.cloudservice.v1.CreateServiceAccountRequest\x1a@.temporal.api.cloud.cloudservice.v1.CreateServiceAccountResponse\"\xd9\x01\x92\x41\xb3\x01\n\x10Service Accounts\x12\x18\x43reate a service account\x1a\x32\x43reates a new service account for automated access\"Q\n\x1eService Accounts documentation\x12/https://docs.temporal.io/cloud/service-accounts\x82\xd3\xe4\x93\x02\x1c\"\x17/cloud/service-accounts:\x01*\x12\x8c\x03\n\x11GetServiceAccount\x12<.temporal.api.cloud.cloudservice.v1.GetServiceAccountRequest\x1a=.temporal.api.cloud.cloudservice.v1.GetServiceAccountResponse\"\xf9\x01\x92\x41\xc1\x01\n\x10Service Accounts\x12\x1bGet service account details\x1a=Returns detailed information about a specific service account\"Q\n\x1eService Accounts documentation\x12/https://docs.temporal.io/cloud/service-accounts\x82\xd3\xe4\x93\x02.\x12,/cloud/service-accounts/{service_account_id}\x12\xf0\x02\n\x12GetServiceAccounts\x12=.temporal.api.cloud.cloudservice.v1.GetServiceAccountsRequest\x1a>.temporal.api.cloud.cloudservice.v1.GetServiceAccountsResponse\"\xda\x01\x92\x41\xb7\x01\n\x10Service Accounts\x12\x19List all service accounts\x1a\x35Returns a list of all service accounts in the account\"Q\n\x1eService Accounts documentation\x12/https://docs.temporal.io/cloud/service-accounts\x82\xd3\xe4\x93\x02\x19\x12\x17/cloud/service-accounts\x12\x88\x03\n\x14UpdateServiceAccount\x12?.temporal.api.cloud.cloudservice.v1.UpdateServiceAccountRequest\x1a@.temporal.api.cloud.cloudservice.v1.UpdateServiceAccountResponse\"\xec\x01\x92\x41\xb1\x01\n\x10Service Accounts\x12\x18Update a service account\x1a\x30Updates an existing service account\'s properties\"Q\n\x1eService Accounts documentation\x12/https://docs.temporal.io/cloud/service-accounts\x82\xd3\xe4\x93\x02\x31\",/cloud/service-accounts/{service_account_id}:\x01*\x12\xe9\x03\n SetServiceAccountNamespaceAccess\x12K.temporal.api.cloud.cloudservice.v1.SetServiceAccountNamespaceAccessRequest\x1aL.temporal.api.cloud.cloudservice.v1.SetServiceAccountNamespaceAccessResponse\"\xa9\x02\x92\x41\xd0\x01\n\x10Service Accounts\x12$Set service account namespace access\x1a\x43\x43onfigures a service account\'s permissions for a specific namespace\"Q\n\x1eService Accounts documentation\x12/https://docs.temporal.io/cloud/service-accounts\x82\xd3\xe4\x93\x02O\"J/cloud/namespaces/{namespace}/service-accounts/{service_account_id}/access:\x01*\x12\xff\x02\n\x14\x44\x65leteServiceAccount\x12?.temporal.api.cloud.cloudservice.v1.DeleteServiceAccountRequest\x1a@.temporal.api.cloud.cloudservice.v1.DeleteServiceAccountResponse\"\xe3\x01\x92\x41\xab\x01\n\x10Service Accounts\x12\x18\x44\x65lete a service account\x1a*Removes a service account from the account\"Q\n\x1eService Accounts documentation\x12/https://docs.temporal.io/cloud/service-accounts\x82\xd3\xe4\x93\x02.*,/cloud/service-accounts/{service_account_id}\x12\xcb\x01\n\x08GetUsage\x12\x33.temporal.api.cloud.cloudservice.v1.GetUsageRequest\x1a\x34.temporal.api.cloud.cloudservice.v1.GetUsageResponse\"T\x92\x41=\n\x07\x41\x63\x63ount\x12\x0eGet usage data\x1a Get usage data across namespacesX\x01\x82\xd3\xe4\x93\x02\x0e\x12\x0c/cloud/usage\x12\xb0\x02\n\nGetAccount\x12\x35.temporal.api.cloud.cloudservice.v1.GetAccountRequest\x1a\x36.temporal.api.cloud.cloudservice.v1.GetAccountResponse\"\xb2\x01\x92\x41\x98\x01\n\x07\x41\x63\x63ount\x12\x13Get account details\x1a.Returns detailed information about the account\"H\n\x15\x42illing documentation\x12/https://docs.temporal.io/cloud/billing-and-cost\x82\xd3\xe4\x93\x02\x10\x12\x0e/cloud/account\x12\xbb\x02\n\rUpdateAccount\x12\x38.temporal.api.cloud.cloudservice.v1.UpdateAccountRequest\x1a\x39.temporal.api.cloud.cloudservice.v1.UpdateAccountResponse\"\xb4\x01\x92\x41\x97\x01\n\x07\x41\x63\x63ount\x12\x16Update account details\x1a*Updates account configuration and settings\"H\n\x15\x42illing documentation\x12/https://docs.temporal.io/cloud/billing-and-cost\x82\xd3\xe4\x93\x02\x13\"\x0e/cloud/account:\x01*\x12\xf3\x02\n\x19\x43reateNamespaceExportSink\x12\x44.temporal.api.cloud.cloudservice.v1.CreateNamespaceExportSinkRequest\x1a\x45.temporal.api.cloud.cloudservice.v1.CreateNamespaceExportSinkResponse\"\xc8\x01\x92\x41\x8f\x01\n\x06\x45xport\x12\x1a\x43reate history export sink\x1a*Creates a new workflow history export sink\"=\n\x14\x45xport documentation\x12%https://docs.temporal.io/cloud/export\x82\xd3\xe4\x93\x02/\"*/cloud/namespaces/{namespace}/export-sinks:\x01*\x12\x8c\x03\n\x16GetNamespaceExportSink\x12\x41.temporal.api.cloud.cloudservice.v1.GetNamespaceExportSinkRequest\x1a\x42.temporal.api.cloud.cloudservice.v1.GetNamespaceExportSinkResponse\"\xea\x01\x92\x41\xad\x01\n\x06\x45xport\x12\x18Get history sink details\x1aJReturns detailed information about a specific workflow history export sink\"=\n\x14\x45xport documentation\x12%https://docs.temporal.io/cloud/export\x82\xd3\xe4\x93\x02\x33\x12\x31/cloud/namespaces/{namespace}/export-sinks/{name}\x12\x82\x03\n\x17GetNamespaceExportSinks\x12\x42.temporal.api.cloud.cloudservice.v1.GetNamespaceExportSinksRequest\x1a\x43.temporal.api.cloud.cloudservice.v1.GetNamespaceExportSinksResponse\"\xdd\x01\x92\x41\xa7\x01\n\x06\x45xport\x12\x19List history export sinks\x1a\x43Returns a list of all workflow history export sinks for a namespace\"=\n\x14\x45xport documentation\x12%https://docs.temporal.io/cloud/export\x82\xd3\xe4\x93\x02,\x12*/cloud/namespaces/{namespace}/export-sinks\x12\x95\x03\n\x19UpdateNamespaceExportSink\x12\x44.temporal.api.cloud.cloudservice.v1.UpdateNamespaceExportSinkRequest\x1a\x45.temporal.api.cloud.cloudservice.v1.UpdateNamespaceExportSinkResponse\"\xea\x01\x92\x41\xa5\x01\n\x06\x45xport\x12\x1aUpdate history export sink\x1a@Updates an existing workflow history export sink\'s configuration\"=\n\x14\x45xport documentation\x12%https://docs.temporal.io/cloud/export\x82\xd3\xe4\x93\x02;\"6/cloud/namespaces/{namespace}/export-sinks/{spec.name}:\x01*\x12\x84\x03\n\x19\x44\x65leteNamespaceExportSink\x12\x44.temporal.api.cloud.cloudservice.v1.DeleteNamespaceExportSinkRequest\x1a\x45.temporal.api.cloud.cloudservice.v1.DeleteNamespaceExportSinkResponse\"\xd9\x01\x92\x41\x9c\x01\n\x06\x45xport\x12\x1a\x44\x65lete history export sink\x1a\x37Removes a workflow history export sink from a namespace\"=\n\x14\x45xport documentation\x12%https://docs.temporal.io/cloud/export\x82\xd3\xe4\x93\x02\x33*1/cloud/namespaces/{namespace}/export-sinks/{name}\x12\xc9\x03\n\x1bValidateNamespaceExportSink\x12\x46.temporal.api.cloud.cloudservice.v1.ValidateNamespaceExportSinkRequest\x1aG.temporal.api.cloud.cloudservice.v1.ValidateNamespaceExportSinkResponse\"\x98\x02\x92\x41\xd7\x01\n\x06\x45xport\x12*Validate history export sink configuration\x1a\x62Tests workflow history export sink configuration by delivering a test file to verify accessibility\"=\n\x14\x45xport documentation\x12%https://docs.temporal.io/cloud/export\x82\xd3\xe4\x93\x02\x37\"2/cloud/namespaces/{namespace}/export-sink-validate:\x01*\x12\xfc\x02\n\x13UpdateNamespaceTags\x12>.temporal.api.cloud.cloudservice.v1.UpdateNamespaceTagsRequest\x1a?.temporal.api.cloud.cloudservice.v1.UpdateNamespaceTagsResponse\"\xe3\x01\x92\x41\xab\x01\n\nNamespaces\x12\x15Update namespace tags\x1a,Updates the tags associated with a namespace\"X\n\x1bNamespace tag documentation\x12\x39https://docs.temporal.io/cloud/namespaces#tag-a-namespace\x82\xd3\xe4\x93\x02.\")/cloud/namespaces/{namespace}/update-tags:\x01*\x12\xff\x02\n\x16\x43reateConnectivityRule\x12\x41.temporal.api.cloud.cloudservice.v1.CreateConnectivityRuleRequest\x1a\x42.temporal.api.cloud.cloudservice.v1.CreateConnectivityRuleResponse\"\xdd\x01\x92\x41\xb5\x01\n\x12\x43onnectivity Rules\x12\x18\x43reate connectivity rule\x1a:Creates a new connectivity rule for network access control\"I\n\x1a\x43onnectivity documentation\x12+https://docs.temporal.io/cloud/connectivity\x82\xd3\xe4\x93\x02\x1e\"\x19/cloud/connectivity-rules:\x01*\x12\x94\x03\n\x13GetConnectivityRule\x12>.temporal.api.cloud.cloudservice.v1.GetConnectivityRuleRequest\x1a?.temporal.api.cloud.cloudservice.v1.GetConnectivityRuleResponse\"\xfb\x01\x92\x41\xbf\x01\n\x12\x43onnectivity Rules\x12\x1dGet connectivity rule details\x1a?Returns detailed information about a specific connectivity rule\"I\n\x1a\x43onnectivity documentation\x12+https://docs.temporal.io/cloud/connectivity\x82\xd3\xe4\x93\x02\x32\x12\x30/cloud/connectivity-rules/{connectivity_rule_id}\x12\xf6\x02\n\x14GetConnectivityRules\x12?.temporal.api.cloud.cloudservice.v1.GetConnectivityRulesRequest\x1a@.temporal.api.cloud.cloudservice.v1.GetConnectivityRulesResponse\"\xda\x01\x92\x41\xb5\x01\n\x12\x43onnectivity Rules\x12\x1bList all connectivity rules\x1a\x37Returns a list of all connectivity rules in the account\"I\n\x1a\x43onnectivity documentation\x12+https://docs.temporal.io/cloud/connectivity\x82\xd3\xe4\x93\x02\x1b\x12\x19/cloud/connectivity-rules\x12\x85\x03\n\x16\x44\x65leteConnectivityRule\x12\x41.temporal.api.cloud.cloudservice.v1.DeleteConnectivityRuleRequest\x1a\x42.temporal.api.cloud.cloudservice.v1.DeleteConnectivityRuleResponse\"\xe3\x01\x92\x41\xa7\x01\n\x12\x43onnectivity Rules\x12\x18\x44\x65lete connectivity rule\x1a,Removes a connectivity rule from the account\"I\n\x1a\x43onnectivity documentation\x12+https://docs.temporal.io/cloud/connectivity\x82\xd3\xe4\x93\x02\x32*0/cloud/connectivity-rules/{connectivity_rule_id}\x12\xe2\x02\n\x0cGetAuditLogs\x12\x37.temporal.api.cloud.cloudservice.v1.GetAuditLogsRequest\x1a\x38.temporal.api.cloud.cloudservice.v1.GetAuditLogsResponse\"\xde\x01\x92\x41\xc1\x01\n\x07\x41\x63\x63ount\x12\x0eGet audit logs\x1aYReturns a paginated list of audit logs for the account, optionally filtered by time range\"K\n\x1b\x41udit logging documentation\x12,https://docs.temporal.io/cloud/audit-logging\x82\xd3\xe4\x93\x02\x13\x12\x11/cloud/audit-logs\x12\xb4\x04\n\x1bValidateAccountAuditLogSink\x12\x46.temporal.api.cloud.cloudservice.v1.ValidateAccountAuditLogSinkRequest\x1aG.temporal.api.cloud.cloudservice.v1.ValidateAccountAuditLogSinkResponse\"\x83\x03\x92\x41\xd6\x02\n\x07\x41\x63\x63ount\x12\x17Validate audit log sink\x1a\xe4\x01Validate customer audit log sink is accessible from Temporal\'s workflow by delivering an empty file to the specified sink. The operation verifies that the sink is correctly configured, accessible and ready to receive audit logs.\"K\n\x1b\x41udit logging documentation\x12,https://docs.temporal.io/cloud/audit-logging\x82\xd3\xe4\x93\x02#\"\x1e/cloud/audit-log-sink-validate:\x01*\x12\xf4\x02\n\x19\x43reateAccountAuditLogSink\x12\x44.temporal.api.cloud.cloudservice.v1.CreateAccountAuditLogSinkRequest\x1a\x45.temporal.api.cloud.cloudservice.v1.CreateAccountAuditLogSinkResponse\"\xc9\x01\x92\x41\xa4\x01\n\x07\x41\x63\x63ount\x12\x15\x43reate audit log sink\x1a\x35\x43reates a new audit log sink for exporting audit logs\"K\n\x1b\x41udit logging documentation\x12,https://docs.temporal.io/cloud/audit-logging\x82\xd3\xe4\x93\x02\x1b\"\x16/cloud/audit-log-sinks:\x01*\x12\xfb\x02\n\x16GetAccountAuditLogSink\x12\x41.temporal.api.cloud.cloudservice.v1.GetAccountAuditLogSinkRequest\x1a\x42.temporal.api.cloud.cloudservice.v1.GetAccountAuditLogSinkResponse\"\xd9\x01\x92\x41\xb0\x01\n\x07\x41\x63\x63ount\x12\x1aGet audit log sink details\x1a<Returns detailed information about a specific audit log sink\"K\n\x1b\x41udit logging documentation\x12,https://docs.temporal.io/cloud/audit-logging\x82\xd3\xe4\x93\x02\x1f\x12\x1d/cloud/audit-log-sinks/{name}\x12\xea\x02\n\x17GetAccountAuditLogSinks\x12\x42.temporal.api.cloud.cloudservice.v1.GetAccountAuditLogSinksRequest\x1a\x43.temporal.api.cloud.cloudservice.v1.GetAccountAuditLogSinksResponse\"\xc5\x01\x92\x41\xa3\x01\n\x07\x41\x63\x63ount\x12\x14List audit log sinks\x1a\x35Returns a list of all audit log sinks for the account\"K\n\x1b\x41udit logging documentation\x12,https://docs.temporal.io/cloud/audit-logging\x82\xd3\xe4\x93\x02\x18\x12\x16/cloud/audit-log-sinks\x12\xfd\x02\n\x19UpdateAccountAuditLogSink\x12\x44.temporal.api.cloud.cloudservice.v1.UpdateAccountAuditLogSinkRequest\x1a\x45.temporal.api.cloud.cloudservice.v1.UpdateAccountAuditLogSinkResponse\"\xd2\x01\x92\x41\xa1\x01\n\x07\x41\x63\x63ount\x12\x15Update audit log sink\x1a\x32Updates an existing audit log sink\'s configuration\"K\n\x1b\x41udit logging documentation\x12,https://docs.temporal.io/cloud/audit-logging\x82\xd3\xe4\x93\x02\'\"\"/cloud/audit-log-sinks/{spec.name}:\x01*\x12\xed\x02\n\x19\x44\x65leteAccountAuditLogSink\x12\x44.temporal.api.cloud.cloudservice.v1.DeleteAccountAuditLogSinkRequest\x1a\x45.temporal.api.cloud.cloudservice.v1.DeleteAccountAuditLogSinkResponse\"\xc2\x01\x92\x41\x99\x01\n\x07\x41\x63\x63ount\x12\x15\x44\x65lete audit log sink\x1a*Removes an audit log sink from the account\"K\n\x1b\x41udit logging documentation\x12,https://docs.temporal.io/cloud/audit-logging\x82\xd3\xe4\x93\x02\x1f*\x1d/cloud/audit-log-sinks/{name}\x12\x9a\x04\n\x18GetNamespaceCapacityInfo\x12\x43.temporal.api.cloud.cloudservice.v1.GetNamespaceCapacityInfoRequest\x1a\x44.temporal.api.cloud.cloudservice.v1.GetNamespaceCapacityInfoResponse\"\xf2\x02\x92\x41\xbb\x02\n\nNamespaces\x12\"Get namespace capacity information\x1a\xad\x01Returns capacity information for a namespace. This includes provisioned capacity options, on-demand limits, and 7 day historical APS statistics useful for capacity planning.\"Y\n(Capacity modes information documentation\x12-https://docs.temporal.io/cloud/capacity-modes\x82\xd3\xe4\x93\x02-\x12+/cloud/namespaces/{namespace}/capacity-info\x12\xda\x02\n\x13\x43reateBillingReport\x12>.temporal.api.cloud.cloudservice.v1.CreateBillingReportRequest\x1a?.temporal.api.cloud.cloudservice.v1.CreateBillingReportResponse\"\xc1\x01\x92\x41\x9c\x01\n\x07\x41\x63\x63ount\x12\x17\x43reate a billing report\x1a(Creates a billing report for the account\"N\n\x1c\x42illing report documentation\x12.https://docs.temporal.io/cloud/billing-reports\x82\xd3\xe4\x93\x02\x1b\"\x16/cloud/billing-reports:\x01*\x12\xe6\x02\n\x10GetBillingReport\x12;.temporal.api.cloud.cloudservice.v1.GetBillingReportRequest\x1a<.temporal.api.cloud.cloudservice.v1.GetBillingReportResponse\"\xd6\x01\x92\x41\xa0\x01\n\x07\x41\x63\x63ount\x12\x14Get a billing report\x1a/Gets an existing billing report for the account\"N\n\x1c\x42illing report documentation\x12.https://docs.temporal.io/cloud/billing-reports\x82\xd3\xe4\x93\x02,\x12*/cloud/billing-reports/{billing_report_id}B\x86\x15\n%io.temporal.api.cloud.cloudservice.v1B\x0cServiceProtoP\x01Z5go.temporal.io/api/cloud/cloudservice/v1;cloudservice\xaa\x02$Temporalio.Api.Cloud.CloudService.V1\xea\x02(Temporalio::Api::Cloud::CloudService::V1\x92\x41\xc2\x13\x12\xe0\r\n\x16Temporal Cloud Ops API\x12\x96\x0cProgrammatic access to manage Temporal Cloud control plane resources including namespaces, users, service accounts, and more.\n\n## Authentication\n\nAll API requests require authentication using an API Key. Include your API key in the `Authorization` header using the Bearer scheme:\n\n```\nAuthorization: Bearer YOUR_API_KEY\n```\n\nAPI keys can be created and managed through the [API Keys endpoints](#tag/API-Keys) or via the Temporal Cloud UI. For more information, see [API Keys Documentation](https://docs.temporal.io/cloud/api-keys).\n\n## Authorization\n\nThe API uses Role-Based Access Control (RBAC) to manage permissions. Each operation requires specific role-based permissions in addition to a valid API key.\n\n### Account-Level Roles\n\n- **Account Owner** - Full account administration access\n- **Account Admin** - Manage namespaces, users, and service accounts  \n- **Account Developer** - Create namespaces and manage Nexus endpoints\n- **Finance Admin** - View usage and billing information\n- **Account Read** - Read-only access to account resources\n\n### Namespace-Level Roles\n\n- **Namespace Admin** - Full access to namespace configuration and data\n- **Namespace Write** - Execute workflows and modify workflow data\n- **Namespace Read** - Read-only access to namespace data\n\nNamespace-level permissions are scoped to specific namespaces. A user or service account may have different permission levels across different namespaces.\n\nFor detailed information about roles and permissions, see [Access Control Documentation](https://docs.temporal.io/cloud/users).2\x03\x31.0:\xa7\x01\n\x06x-logo\x12\x9c\x01*\x99\x01\n\x96\x01\n\x03url\x12\x8e\x01\x1a\x8b\x01https://images.ctfassets.net/0uuz8ydxyd9p/4YGUnEoCaH9SyoUDhlJkau/e1600205d17eeee3033d926ef06664a9/Temporal_LogoLockup_Horizontal_dark_1.svgj.\n\nNamespaces\x12 Manage Temporal Cloud namespacesj0\n\x05Users\x12\'Manage users and their namespace accessjF\n\x10Service Accounts\x12\x32Manage service accounts and their namespace accessj.\n\x08\x41PI Keys\x12\"Manage API keys for authenticationj1\n\x06Groups\x12\'Manage user groups and group membershipj\x1f\n\x05Nexus\x12\x16Manage Nexus endpointsj\x7f\n\x11High Availability\x12jManage high availability (multi-region, multi-cloud, and same-region replication) namespace configurationsj7\n\x06\x45xport\x12-Manage workflow history export configurationsj7\n\x12\x43onnectivity Rules\x12!Manage network connectivity rulesj\"\n\x07Regions\x12\x17Query available regionsj,\n\x07\x41\x63\x63ount\x12!Manage account settings and usagej*\n\nOperations\x12\x1cQuery async operation statusr>\n\x1cTemporal Cloud Documentation\x12\x1ehttps://docs.temporal.io/cloudb\x06proto3"
