+++ temporalio/lib/temporalio/api/protoc_gen_openapiv2/options/openapiv2.rb
