+++ temporalio/sig/temporalio/api/protoc_gen_openapiv2/options/annotations.rbs
@@ -1,0 +1,23 @@
+# Generated by the protocol buffer compiler.  DO NOT EDIT!
+# This RBS interface is provided for convenience, on a best-effort basis.
+# The library is the definitive source for the API contract; if the RBS file
+# and the library's behavior differ, the library behavior is authoritative.
+# We welcome fixes to change the RBS file to match.
+# source: protoc-gen-openapiv2/options/annotations.proto
+
+module Grpc
+  module Gateway
+    module ProtocGenOpenapiv2
+      module Options
+      end
+    end
+  end
+end
+
+module Google
+  module Protobuf
+    class DescriptorPool
+      def lookup: (::String name) -> ::Google::Protobuf::Descriptor
+    end
+  end
+end
